// src/api/generated/model/commonResponseFindMyPageResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Tudak API
 * 투닥 API 명세서
 * OpenAPI spec version: v1.0.0
 */
import type { FindMyPageResponse } from "./findMyPageResponse";

export interface CommonResponseFindMyPageResponse {
  success?: boolean;
  data?: FindMyPageResponse;
  message?: string;
}